notifier/pyproject.toml
[tool.poetry]
name = "notifier"
version = "0.1.0"
description = ""
authors = ["Matvey <motiio.none@gmail.com>"]
readme = "README.md"

[tool.poetry.group.dev.dependencies]
neovim = "^0.3.1"
ruff = "^0.1.11"

[tool.pyright]
venvPath = "$HOME/Library/Caches/pypoetry/virtualenvs/"
venv = "notifier-AuczI1ej-py3.12"

[tool.mypy]
ignore_missing_imports = true
show_column_numbers = true
show_error_codes = true
pretty = true
show_error_context=true

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.108.0"
websockets = "^12.0"
uvicorn = "^0.25.0"
pydantic-settings = "^2.1.0"
python-jose = "^3.3.0"
aio-pika = "^9.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
